requirements.txt
streamlit
google-cloud-aiplatform
google-cloud-firestore
google-generativeai
st_pages
python-dotenv
ipywidgets